Fix #1460: The kubernetes-client is not optionally depending on bouncycastle.
Fixed my mistake
