+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/Config.java
@@ -670,10 +670,7 @@ public class Config {
         return getKeyAlgorithm(keyInputStream);
       }
     } catch(IOException exception) {
-      LOGGER.info("Failure in determining private key algorithm type, defaulting to RSA ", exception.getMessage());
-    } catch(NoClassDefFoundError noClassDefFoundError) {
-      throw new IllegalStateException("Please make sure org.bouncycastle:bcprov-ext-jdk15on " +
-        "and org.bouncycastle:bcpkix-jdk15on are included in your project's classpath. ", noClassDefFoundError);
+      LOGGER.debug("Failure in determining private key algorithm type, defaulting to RSA ", exception.getMessage());
     }
     return null;
   }
