+++ kubernetes-tests/src/test/resources/ssl/fabric8-ec.cert
