+++ platforms/karaf/features/src/main/resources/feature.xml
@@ -19,9 +19,6 @@
 
   <feature name="kubernetes-client" description="Fabric8 Kubernetes Client" version="${project.version}">
     <bundle dependency='true'>mvn:javax.validation/validation-api/${validation.api.version}</bundle>
-    <bundle dependency='true'>mvn:org.bouncycastle/bcprov-ext-jdk15on/${bouncycastle.version}</bundle>
-    <bundle dependency='true'>mvn:org.bouncycastle/bcpkix-jdk15on/${bouncycastle.version}</bundle>
-
     <bundle dependency='true'>mvn:com.google.code.findbugs/jsr305/${jsr305.version}</bundle>
     <bundle dependency='true'>mvn:com.fasterxml.jackson.core/jackson-core/${jackson.version}</bundle>
     <bundle dependency='true'>mvn:com.fasterxml.jackson.core/jackson-annotations/${jackson.bundle.version}</bundle>
